requirements.txt
pyyaml>=6.0
flask>=3.0
markdown>=3.0
